RESUME_ATS/requirements.txt
streamlit
pdfplumber
python-docx
PyMuPDF
nltk
spacy>=3.7.2
sentence-transformers
scikit-learn
xgboost
joblib
en_core_web_sm>=3.7.0
